Add support for unicode type/literals
This is a very minimal support for unicode strings (not actually
support unicode in the literals right now) which shares all literals
used by the same module as static globals.

I debated between calling it string and calling it unicode, and decided
on unicode because:

(1) Its more explicit about what it actually contains, which is good

(2) The C API calls it PyUnicode so its not inconsistent with that
layer, which is what we are actually interoperating with in this
case.